% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapingDistrict.R
\name{ScrapingDistrict}
\alias{ScrapingDistrict}
\title{Scraping District Information}
\usage{
ScrapingDistrict(MemberWebpage)
}
\arguments{
\item{MemberWebpage}{The html of a politician's webpage obtained through the read_html function}
}
\value{
District information of city council members
}
\description{
Scraping District Information
}
\examples{
url <- "https://ballotpedia.org/Angelique_Ashby"
MemberWebpage <- xml2::read_html(url)
District <- ScrapingDistrict(MemberWebpage); District
"Sacramento City Council District 1"
}
