% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapingParty.R
\name{ScrapingParty}
\alias{ScrapingParty}
\title{Scraping Party Information}
\usage{
ScrapingParty(MemberWebpage)
}
\arguments{
\item{MemberWebpage}{The html of a politician's webpage obtained through the read_html function}
}
\value{
Party information
}
\description{
Scraping Party Information
}
\examples{
url <- "https://ballotpedia.org/Angelique_Ashby"
MemberWebpage <- xml2::read_html(url)
Party <- ScrapingParty(MemberWebpage); Party
"Democratic Party"
}
