% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapol_city.R
\name{scrapol_city}
\alias{scrapol_city}
\title{Scraping Politicians' Party and District Information Across American Cities and Counties}
\usage{
scrapol_city(url, sleep)
}
\arguments{
\item{url}{A vector of urls pointing to politicians' webpages on Ballotpedia}

\item{sleep}{The sleep time between two scraping trials}
}
\value{
A dataframe consists of url, partisanship, and district
}
\description{
Scraping Politicians' Party and District Information Across American Cities and Counties
}
\examples{
url <- c("https://ballotpedia.org/Angelique_Ashby", "https://ballotpedia.org/Connie_Dodson")
scrapol_city(url, sleep=15)
}
